requirements.txt
Django==1.6.4
South==0.8.4
django-braces==1.4.0
djangorestframework==2.3.13
psycopg2==2.5.2
six==1.6.1
wsgiref==0.1.2
